requirements.txt
langchain
streamlit
replicate
python-dotenv